Move batch-insert into execution directory since it's not really a part of query building
