+++ lib/execution/batch-insert.js
@@ -1,7 +1,7 @@
 const chunk = require('lodash/chunk');
 const flatten = require('lodash/flatten');
-const delay = require('./delay');
-const { isNumber } = require('../../util/is');
+const delay = require('./internal/delay');
+const { isNumber } = require('../util/is');
 
 function batchInsert(client, tableName, batch, chunkSize = 1000) {
   let returning = void 0;
