+++ lib/knex-builder/make-knex.js
@@ -5,7 +5,7 @@ const Seeder = require('../migrations/seed/Seeder');
 const FunctionHelper = require('./FunctionHelper');
 const QueryInterface = require('../query/methods');
 const merge = require('lodash/merge');
-const batchInsert = require('./internal/batch-insert');
+const batchInsert = require('../execution/batch-insert');
 
 // Javascript does not officially support "callable objects".  Instead,
 // you must create a regular Function and inject properties/methods
