+++ scripts/stress-test/mysql2-sudden-exit-without-error.js
@@ -3,7 +3,7 @@
  * to dead connection.
  */
 
-const delay = require('../../lib/knex-builder/internal/delay');
+const delay = require('../../lib/execution/internal/delay');
 const toxiproxy = require('toxiproxy-node-client');
 const toxicli = new toxiproxy.Toxiproxy('http://localhost:8474');
 const rp = require('request-promise-native');
