+++ test/integration/query/additional.js
@@ -5,7 +5,7 @@ const { expect } = require('chai');
 
 const Knex = require('../../../knex');
 const _ = require('lodash');
-const delay = require('../../../lib/knex-builder/internal/delay');
+const delay = require('../../../lib/execution/internal/delay');
 const {
   isPostgreSQL,
   isOracle,
